Include generation of slab bridge interim CF
The Canonical Form (CF) reduction rules currently do not have the
rules to reduce an Irreducible Element (IE) model to the final version
of a CF. As such, an interim version of the CF is required for fair
comparisons between none ML and ML similarity metrics. The generated
IE models include the same as the CF versions, however the interim
version includes an additional node at the begining of the deck and the
end of the deck.